moved new res_finder.py file to utilities
